# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cross-Domain Recommendation Engine API",
    "description": "API documentation for the Cross-Domain Recommendation Engine. This spec should be updated as endpoints change.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Authentication using Clerk. Format: Bearer {token}"
      }
    },
...
  },
  "security": [
    { "clerkAuth": [] }
  ],
  "paths": {
...
    "/v1/recommendations": {
      "post": {
        "summary": "Get cross-domain recommendations",
        "description": "Generate personalized recommendations across multiple domains based on input entities",
        "tags": ["Recommendations"],
        "security": [{ "clerkAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["entities"],
                "properties": {
                  "entities": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Radiohead"
                        },
                        "type": {
                          "type": "string",
                          "example": "artist"
                        }
                      }
                    }
                  },
                  "limit": {
                    "type": "integer",
                    "default": 5,
                    "example": 5
                  },
                  "includeExplanations": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with recommendations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "input": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "recommendations": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}